feat: move model migrate and initModel to database (#9796)
(cherry picked from commit c8292c6c1688faf7fe29cc4e137856a8d4032373)
