+++ packages/insomnia/src/insomnia-data/node-src/database/database-nedb.ts
@@ -24,10 +24,11 @@ import type {
   Workspace,
   WorkspaceMeta,
 } from '~/insomnia-data';
+import type { AllTypes, BaseModel } from '~/models';
 import { mustGetModel } from '~/models';
-import type { AllTypes, BaseModel } from '~/models/index';
-import * as models from '~/models/index';
+import * as models from '~/models';
 
+import { initModel } from './init-model';
 import { repairDatabase } from './repair-database';
 
 const getTempPath = (name: string) => {
@@ -73,7 +74,7 @@ export const createNedbDatabase = <O = initOptions>(
     },
 
     docCreate: async <T extends BaseModel>(type: AllTypes, ...patches: Partial<T>[]) => {
-      const doc = await models.initModel<T>(
+      const doc = await initModel<T>(
         type,
         ...patches,
         // Fields that the user can't touch
@@ -86,7 +87,7 @@ export const createNedbDatabase = <O = initOptions>(
 
     docUpdate: async <T extends BaseModel>(originalDoc: T, ...patches: Partial<T>[]) => {
       // No need to re-initialize the model during update; originalDoc will be in a valid state by virtue of loading
-      const doc = await models.initModel<T>(
+      const doc = await initModel<T>(
         originalDoc.type,
         originalDoc,
 
@@ -161,7 +162,7 @@ export const createNedbDatabase = <O = initOptions>(
       if (doc === null) {
         return undefined;
       }
-      return models.initModel<T>(type, doc);
+      return initModel<T>(type, doc);
     },
     /** find documents matching query */
     find: async function <T extends BaseModel>(
@@ -178,7 +179,7 @@ export const createNedbDatabase = <O = initOptions>(
       // TODO: create a db init phase for migrations rather than doing it on every find.
       const migrated = [];
       for (const rawDoc of docs) {
-        migrated.push(await models.initModel<T>(type, rawDoc));
+        migrated.push(await initModel<T>(type, rawDoc));
       }
       return migrated;
     },
@@ -379,7 +380,7 @@ export const createNedbDatabase = <O = initOptions>(
     },
 
     insert: async function <T extends BaseModel>(doc: T) {
-      const docWithDefaults = await models.initModel<T>(doc.type, doc);
+      const docWithDefaults = await initModel<T>(doc.type, doc);
       const newDoc = await nedbBucket[doc.type].insertAsync(docWithDefaults);
       notifyOfChange('insert', newDoc);
       return newDoc;
@@ -449,7 +450,7 @@ export const createNedbDatabase = <O = initOptions>(
     },
 
     update: async function <T extends BaseModel>(doc: T, patches: Partial<T>[] = []) {
-      const docWithDefaults = await models.initModel<T>(doc.type, doc);
+      const docWithDefaults = await initModel<T>(doc.type, doc);
       await nedbBucket[doc.type].updateAsync({ _id: docWithDefaults._id }, docWithDefaults, { upsert: true });
       notifyOfChange('update', docWithDefaults, patches);
       return docWithDefaults;
