+++ packages/insomnia/src/insomnia-data/node-src/database/init-model/response.ts
@@ -1,0 +1,18 @@
+import type { Response } from '~/insomnia-data';
+
+export function migrate(doc: Response) {
+  try {
+    return migrateBodyCompression(doc);
+  } catch (e) {
+    console.log('[db] Error during response migration', e);
+    throw e;
+  }
+}
+
+function migrateBodyCompression(doc: Response) {
+  if (doc.bodyCompression === '__NEEDS_MIGRATION__') {
+    doc.bodyCompression = 'zip';
+  }
+
+  return doc;
+}
