+++ packages/insomnia/src/insomnia-data/node-src/database/init-model/settings.ts
@@ -1,0 +1,35 @@
+import * as hotkeys from '~/common/hotkeys';
+import type { KeyboardShortcut } from '~/common/settings';
+import type { Settings } from '~/insomnia-data';
+
+export function migrate(doc: Settings) {
+  try {
+    doc = migrateEnsureHotKeys(doc);
+    return doc;
+  } catch (e) {
+    console.log('[db] Error during settings migration', e);
+    throw e;
+  }
+}
+
+/**
+ * Ensure map is updated when new hotkeys are added
+ */
+function migrateEnsureHotKeys(settings: Settings): Settings {
+  const defaultHotKeyRegistry = hotkeys.newDefaultRegistry();
+
+  // Remove any hotkeys that are no longer in the default registry
+  const hotKeyRegistry = (Object.keys(settings.hotKeyRegistry) as KeyboardShortcut[]).reduce(
+    (newHotKeyRegistry, key) => {
+      if (key in defaultHotKeyRegistry) {
+        newHotKeyRegistry[key] = settings.hotKeyRegistry[key];
+      }
+
+      return newHotKeyRegistry;
+    },
+    {} as Settings['hotKeyRegistry'],
+  );
+
+  settings.hotKeyRegistry = { ...defaultHotKeyRegistry, ...hotKeyRegistry };
+  return settings;
+}
