+++ packages/insomnia/src/insomnia-data/node-src/services/client-certificate.ts
@@ -1,4 +1,5 @@
-import { type ClientCertificate, database as db, models } from '~/insomnia-data';
+import type { ClientCertificate } from '~/insomnia-data';
+import { database as db, models } from '~/insomnia-data';
 
 const { type } = models.clientCertificate;
