+++ packages/insomnia/src/insomnia-data/node-src/services/cloud-credential.ts
@@ -1,4 +1,5 @@
-import { type CloudProviderCredential, type CloudProviderName, database as db, models } from '~/insomnia-data';
+import type { CloudProviderCredential, CloudProviderName } from '~/insomnia-data';
+import { database as db, models } from '~/insomnia-data';
 
 const { type } = models.cloudCredential;
