+++ packages/insomnia/src/insomnia-data/node-src/services/environment.ts
@@ -1,9 +1,7 @@
 import * as crypto from 'node:crypto';
 
 import type { Environment, Project, Workspace } from '~/insomnia-data';
-import { models } from '~/insomnia-data';
-
-import { database as db } from '../../src/database';
+import { database as db, models } from '~/insomnia-data';
 
 const { type, prefix, vaultEnvironmentPath } = models.environment;
 const { EnvironmentKvPairDataType, EnvironmentType } = models.environment;
