+++ packages/insomnia/src/insomnia-data/node-src/services/grpc-request.ts
@@ -1,6 +1,5 @@
-import { database as db } from '../../src/database';
-import { models } from '../../src/models';
-import { type GrpcRequest } from '../../src/models/types';
+import type { GrpcRequest } from '~/insomnia-data';
+import { database as db, models } from '~/insomnia-data';
 
 const { type, name } = models.grpcRequest;
