+++ packages/insomnia/src/insomnia-data/node-src/services/mock-route.ts
@@ -1,6 +1,5 @@
-import { database as db } from '../../src/database';
-import { models } from '../../src/models';
-import { type MockRoute } from '../../src/models/types';
+import type { MockRoute } from '~/insomnia-data';
+import { database as db, models } from '~/insomnia-data';
 
 const { type } = models.mockRoute;
