+++ packages/insomnia/src/insomnia-data/node-src/services/o-auth-2-token.ts
@@ -1,4 +1,5 @@
-import { database as db, models, type OAuth2Token } from '~/insomnia-data';
+import type { OAuth2Token } from '~/insomnia-data';
+import { database as db, models } from '~/insomnia-data';
 
 const { type } = models.oAuth2Token;
