+++ packages/insomnia/src/insomnia-data/node-src/services/plugin-data.ts
@@ -1,4 +1,5 @@
-import { database as db, models, type PluginData } from '~/insomnia-data';
+import type { PluginData } from '~/insomnia-data';
+import { database as db, models } from '~/insomnia-data';
 
 const { type } = models.pluginData;
