+++ packages/insomnia/src/insomnia-data/node-src/services/project.ts
@@ -1,6 +1,5 @@
-import { database as db } from '../../src/database';
-import { models } from '../../src/models';
-import type { Project } from '../../src/models/types';
+import type { Project } from '~/insomnia-data';
+import { database as db, models } from '~/insomnia-data';
 
 const { type } = models.project;
