+++ packages/insomnia/src/insomnia-data/node-src/services/proto-directory.ts
@@ -1,4 +1,5 @@
-import { database as db, models, type ProtoDirectory } from '~/insomnia-data';
+import type { ProtoDirectory } from '~/insomnia-data';
+import { database as db, models } from '~/insomnia-data';
 
 const { type } = models.protoDirectory;
