+++ packages/insomnia/src/insomnia-data/node-src/services/proto-file.ts
@@ -1,4 +1,5 @@
-import { database as db, models, type ProtoFile } from '~/insomnia-data';
+import type { ProtoFile } from '~/insomnia-data';
+import { database as db, models } from '~/insomnia-data';
 
 const { type } = models.protoFile;
