+++ packages/insomnia/src/insomnia-data/node-src/services/runner-test-result.ts
@@ -1,4 +1,5 @@
-import { database as db, models, type RunnerTestResult } from '~/insomnia-data';
+import type { RunnerTestResult } from '~/insomnia-data';
+import { database as db, models } from '~/insomnia-data';
 
 const { type } = models.runnerTestResult;
