+++ packages/insomnia/src/insomnia-data/node-src/services/socket-io-payload.ts
@@ -1,6 +1,5 @@
-import { database } from '../../src/database';
-import { models } from '../../src/models';
-import { type SocketIOPayload } from '../../src/models/types';
+import type { SocketIOPayload } from '~/insomnia-data';
+import { database, models } from '~/insomnia-data';
 
 const { type, name } = models.socketIOPayload;
