+++ packages/insomnia/src/insomnia-data/node-src/services/unit-test-result.ts
@@ -1,6 +1,5 @@
-import { database as db } from '../../src/database';
-import { models } from '../../src/models';
-import type { UnitTestResult } from '../../src/models/types';
+import type { UnitTestResult } from '~/insomnia-data';
+import { database as db, models } from '~/insomnia-data';
 
 const { type } = models.unitTestResult;
