+++ packages/insomnia/src/insomnia-data/node-src/services/user-session.ts
@@ -1,4 +1,5 @@
-import { database as db, models, type UserSession } from '~/insomnia-data';
+import type { UserSession } from '~/insomnia-data';
+import { database as db, models } from '~/insomnia-data';
 
 const { type } = models.userSession;
