+++ packages/insomnia/src/insomnia-data/src/models/cookie-jar.ts
@@ -1,5 +1,3 @@
-import { v4 as uuidv4 } from 'uuid';
-
 import type { BaseModel } from '~/models/types';
 
 export const name = 'Cookie Jar';
@@ -43,25 +41,4 @@ export function init() {
     name: 'Default Jar',
     cookies: [],
   };
-}
-
-/** Ensure every cookie has an ID property */
-function migrateCookieId(cookieJar: CookieJar) {
-  for (const cookie of cookieJar.cookies) {
-    if (!cookie.id) {
-      cookie.id = uuidv4();
-    }
-  }
-
-  return cookieJar;
-}
-
-export function migrate(doc: CookieJar) {
-  try {
-    doc = migrateCookieId(doc);
-    return doc;
-  } catch (e) {
-    console.log('[db] Error during cookie jar migration', e);
-    throw e;
-  }
 }
