+++ packages/insomnia/src/insomnia-data/src/models/settings.ts
@@ -1,6 +1,6 @@
 import { getAppDefaultDarkTheme, getAppDefaultLightTheme, getAppDefaultTheme } from '~/common/constants';
 import * as hotkeys from '~/common/hotkeys';
-import { HttpVersions, type KeyboardShortcut, type Settings as BaseSettings, UpdateChannel } from '~/common/settings';
+import { HttpVersions, type Settings as BaseSettings, UpdateChannel } from '~/common/settings';
 import type { BaseModel } from '~/models/types';
 
 export type Settings = BaseModel & BaseSettings;
@@ -76,36 +76,4 @@ export function init(): BaseSettings {
     vaultSecretCacheDuration: 30,
     dataFolders: [],
   };
-}
-
-export function migrate(doc: Settings) {
-  try {
-    doc = migrateEnsureHotKeys(doc);
-    return doc;
-  } catch (e) {
-    console.log('[db] Error during settings migration', e);
-    throw e;
-  }
-}
-
-/**
- * Ensure map is updated when new hotkeys are added
- */
-function migrateEnsureHotKeys(settings: Settings): Settings {
-  const defaultHotKeyRegistry = hotkeys.newDefaultRegistry();
-
-  // Remove any hotkeys that are no longer in the default registry
-  const hotKeyRegistry = (Object.keys(settings.hotKeyRegistry) as KeyboardShortcut[]).reduce(
-    (newHotKeyRegistry, key) => {
-      if (key in defaultHotKeyRegistry) {
-        newHotKeyRegistry[key] = settings.hotKeyRegistry[key];
-      }
-
-      return newHotKeyRegistry;
-    },
-    {} as Settings['hotKeyRegistry'],
-  );
-
-  settings.hotKeyRegistry = { ...defaultHotKeyRegistry, ...hotKeyRegistry };
-  return settings;
 }
