+++ packages/insomnia/src/plugins/context/__tests__/response.test.ts
@@ -4,7 +4,8 @@ import path from 'node:path';
 
 import { describe, expect, it } from 'vitest';
 
-import * as models from '../../../models/index';
+import { services } from '~/insomnia-data';
+
 import * as plugin from '../response';
 
 describe('init()', () => {
@@ -35,7 +36,7 @@ describe('response.*', () => {
   it('works for basic and full response', async () => {
     const bodyPath = path.join(tmpdir(), 'response.zip');
     fs.writeFileSync(bodyPath, Buffer.from('Hello World!'));
-    const response = await models.initModel(models.response.type, {
+    const response = await services.response.create({
       bodyPath,
       bodyCompression: null,
       parentId: 'req_1',
